Extract LogTester in a separate module
