+++ plugin-api-test-utils/src/main/java/org/sonar/api/testutils/log/LogTester.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.api.utils.log;
+package org.sonar.api.testutils.log;
 
 import org.junit.rules.TestRule;
 import org.junit.runner.Description;
