+++ plugin-api/build.gradle
@@ -12,15 +12,14 @@ dependencies {
   compileOnly 'ch.qos.logback:logback-classic'
   compileOnly 'com.google.code.findbugs:jsr305'
   compileOnly 'javax.servlet:javax.servlet-api'
-  compileOnly 'junit:junit'
-  // Used by LogTesterJUnit5
-  compileOnly 'org.junit.jupiter:junit-jupiter-api'
   compileOnly 'org.slf4j:slf4j-api'
 
+  testImplementation 'junit:junit'
   testImplementation 'com.google.guava:guava'
   testImplementation 'com.tngtech.java:junit-dataprovider'
   testImplementation 'org.assertj:assertj-core'
   testImplementation 'org.mockito:mockito-core'
+  testImplementation project(':plugin-api-test-utils')
 }
 
 configurations {
