+++ plugin-api/src/test/java/org/sonar/api/utils/log/ConsoleLoggerTest.java
@@ -23,6 +23,7 @@ import java.io.PrintStream;
 import org.junit.Rule;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
+import org.sonar.api.testutils.log.LogTester;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.anyString;
