+++ plugin-api/src/test/java/org/sonar/api/utils/log/LogbackLoggerTest.java
@@ -23,6 +23,7 @@ import ch.qos.logback.classic.Level;
 import org.junit.Rule;
 import org.junit.Test;
 import org.slf4j.LoggerFactory;
+import org.sonar.api.testutils.log.LogTester;
 
 import static org.assertj.core.api.Assertions.assertThat;
