+++ plugin-api/src/test/java/org/sonar/api/utils/log/ProfilerTest.java
@@ -21,6 +21,7 @@ package org.sonar.api.utils.log;
 
 import org.junit.Rule;
 import org.junit.Test;
+import org.sonar.api.testutils.log.LogTester;
 
 import static org.assertj.core.api.Assertions.assertThat;
